Wire up the inference side of the dialogues
Move the responsabilities of the old DialoguePolicy class to
InferenceSentenceGenerator, which initializes the SentenceGenerator
on-demand based on the skills currently in the context.

Introduce a InferenceTimeDialogue, which is a DialogueHandler
for ThingTalk dialogues but also bridges to DialogueInterface
(dialogue PL).

Refactor non-standard dialogue code (interactive-annotate, simulate-dialogs)
to also use InferenceTimeDialogue as a dialogue handler.

Notifications are still WIP.
