+++ lib/dialogue-runtime/dialogue-loop.ts
@@ -41,10 +41,13 @@ import AppExecutor from '../engine/apps/app_executor';
 import DeviceInterfaceMapper from '../engine/devices/device_interface_mapper';
 
 import ExecutionDialogueAgent from '../thingtalk-dialogues/execution_dialogue_agent';
-import ThingTalkDialogueHandler from '../thingtalk-dialogues/thingtalk-dialogue-handler';
+import { InferenceTimeDialogue as ThingTalkDialogueHandler } from '../thingtalk-dialogues/inference-time-dialogue';
+
 import FAQDialogueHandler from './handlers/faq';
 import ThingpediaDialogueHandler from './handlers/3rdparty';
 import DeviceView from '../engine/devices/device_view';
+import { LogLevel } from '../sentence-generator/runtime';
+
 
 export enum CommandAnalysisType {
     // special commands - these are generated by the exact matcher, or
@@ -95,7 +98,7 @@ export interface DialogueHandler<AnalysisType extends CommandAnalysisResult, Sta
 
     analyzeCommand(command : UserInput) : Promise<AnalysisType>;
     getReply(command : AnalysisType) : Promise<ReplyResult>;
-    getFollowUp() : Promise<ReplyResult|null>;
+    //getFollowUp() : Promise<ReplyResult|null>;
 }
 
 export class DialogueLoop {
@@ -129,6 +132,7 @@ export class DialogueLoop {
                 options : {
                     nluServerUrl : string|undefined;
                     nlgServerUrl : string|undefined;
+                    policy : string|undefined;
                     debug : boolean;
                     faqModels : Record<string, {
                         url : string;
@@ -147,7 +151,20 @@ export class DialogueLoop {
         this._nlu = ParserClient.get(options.nluServerUrl || undefined, engine.platform.locale, engine.platform,
             undefined, engine.thingpedia);
         this._nlg = ParserClient.get(options.nlgServerUrl || undefined, engine.platform.locale, engine.platform);
-        this._thingtalkHandler = new ThingTalkDialogueHandler(engine, this, this._agent, this._nlu, this._nlg, options);
+        this._thingtalkHandler = new ThingTalkDialogueHandler({
+            thingpediaClient: engine.thingpedia,
+            schemaRetriever: engine.schemas,
+            locale: engine.platform.locale,
+            timezone: engine.platform.timezone,
+            policy: options.policy,
+            executor: this._agent,
+            nlu: this._nlu,
+            nlg: this._nlg,
+            extraFlags: {},
+            anonymous: conversation.isAnonymous,
+            debug: options.debug ? LogLevel.GENERATION : LogLevel.INFO,
+            rng: Math.random
+        });
         this._faqHandlers = {};
         for (const faq in options.faqModels)
             this._faqHandlers[faq] = new FAQDialogueHandler(this, faq, options.faqModels[faq], { locale: engine.platform.locale });
@@ -361,9 +378,9 @@ export class DialogueLoop {
 
     private async _handleAPICall(call : QueueItem) {
         if (call instanceof QueueItem.Notification)
-            await this._sendAgentReply(await this._thingtalkHandler.showNotification(call.app, call.outputType, call.outputValue));
+            await this._sendAgentReply(await this._thingtalkHandler.showNotification(call.app.program, call.app.name, call.outputType, call.outputValue));
         else if (call instanceof QueueItem.Error)
-            await this._sendAgentReply(await this._thingtalkHandler.showAsyncError(call.app, call.error));
+            await this._sendAgentReply(await this._thingtalkHandler.showAsyncError(call.app.program, call.app.name, call.error));
     }
 
     private async _sendAgentReply(reply : ReplyResult) {
@@ -401,15 +418,6 @@ export class DialogueLoop {
             const reply = await handler.getReply(analysis);
             this.icon = handler.icon;
             await this._sendAgentReply(reply);
-
-            while (this.expecting === null) {
-                const followUp : ReplyResult|null = await handler.getFollowUp();
-                if (followUp === null)
-                    break;
-
-                this.icon = handler.icon;
-                await this._sendAgentReply(followUp);
-            }
 
             // if we're not expecting any more answer from the user,
             // exit this loop
