+++ lib/utils/thingtalk/index.ts
@@ -26,10 +26,10 @@ import { extractConstants, createConstants } from './constants';
 export * from './describe';
 export * from './syntax';
 export * from './dialogue_state_utils';
-import { computePrediction } from './dialogue_state_utils';
 // reexport clean, tokenizeExample from misc-utils
 import { clean, tokenizeExample } from '../misc-utils';
 import { PolicyModule } from '../../thingtalk-dialogues';
+
 export { clean, tokenizeExample };
 
 export {
@@ -102,62 +102,6 @@ export class StateValidator {
         // if and only if
         assert((state.dialogueActParam !== null) === (this._policy.dialogueActs.withParam.has(state.dialogueAct)));
     }
-}
-
-interface DialoguePolicy {
-    handleAnswer(state : Ast.DialogueState, value : Ast.Value) : Promise<Ast.DialogueState|null>;
-}
-
-export async function inputToDialogueState(policy : DialoguePolicy,
-                                           context : Ast.DialogueState|null,
-                                           input : Ast.Input) : Promise<Ast.DialogueState|null> {
-    if (input instanceof Ast.ControlCommand) {
-        if (input.intent instanceof Ast.SpecialControlIntent) {
-            switch (input.intent.type) {
-            case 'yes':
-            case 'no': {
-                if (context === null)
-                    return null;
-                const value = new Ast.BooleanValue(input.intent.type === 'yes');
-                const handled = await policy.handleAnswer(context, value);
-                if (!handled)
-                    return null;
-                return computePrediction(context, handled, 'user');
-            }
-            case 'nevermind':
-                return new Ast.DialogueState(null, 'org.thingpedia.dialogue.transaction', 'cancel', null, []);
-            case 'wakeup':
-                return new Ast.DialogueState(null, 'org.thingpedia.dialogue.transaction', 'greet', null, []);
-            default:
-                return null;
-            }
-        }
-        if (input.intent instanceof Ast.ChoiceControlIntent)
-            return null;
-        if (context === null)
-            return null;
-
-        if (input.intent instanceof Ast.AnswerControlIntent) {
-            const handled = await policy.handleAnswer(context, input.intent.value);
-            if (!handled)
-                return null;
-            return computePrediction(context, handled, 'user');
-        }
-
-        throw new TypeError(`Unrecognized bookkeeping intent`);
-    } else if (input instanceof Ast.Program) {
-        // convert thingtalk programs to dialogue states so we can use "\t" without too much typing
-        const prediction = new Ast.DialogueState(null, 'org.thingpedia.dialogue.transaction', 'execute', null, []);
-        for (const stmt of input.statements) {
-            if (stmt instanceof Ast.Assignment)
-                throw new Error(`Unsupported: assignment statement`);
-            prediction.history.push(new Ast.DialogueHistoryItem(null, stmt, null, 'accepted'));
-        }
-        return prediction;
-    }
-
-    assert(input instanceof Ast.DialogueState);
-    return input;
 }
 
 class UsesParamVisitor extends Ast.NodeVisitor {
